have imperative and symbolic foreach.
